Added a quick start
After building the operator in in Wercker and pushing it to my registry, I realised that I didn't have to do all that. I then watched the youtube video and realised all I had to do was clone, customise and run the actual script except for that one error which required that I have cluster admins rights. While it's great with the in-depth detailed documentation I think the frontpage on the git repo should show easiest and quickest way to try out the operator. Once it works, one can always come back and harden it, customise it further etc. That's when the detailed documentation is great.